requirements.txt
asgiref==3.7.2
blinker==1.7.0
click==8.1.7
dnspython==2.6.1
email_validator==2.1.1
Flask==3.0.2
Flask-Cors==4.0.0
flask-mongoengine @ git+https://github.com/idoshr/flask-mongoengine.git@e244408acf440c4208f7ddcd6e5d819cb472e4da
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
gunicorn==22.0.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mongoengine==0.28.2
packaging==24.0
psycopg2==2.9.9
PyJWT==2.8.0
pymongo==4.6.2
python-dotenv==1.0.1
SQLAlchemy==2.0.29
typing_extensions==4.11.0
Werkzeug==3.0.1
WTForms==3.1.2
